# core-common-sca-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-common-sca-mgmt
  description: Strong Customer Authentication Management Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/sca
    description: Development Environment
  - url: /
    description: Local Development Environment
paths:
  /api/v1/sca/operations/{operationId}:
    get:
      tags:
        - sca-operation-controller
      operationId: getOperation
      parameters:
        - name: operationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SCAOperationDTO'
    put:
      tags:
        - sca-operation-controller
      operationId: updateOperation
      parameters:
        - name: operationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SCAOperationDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SCAOperationDTO'
    delete:
      tags:
        - sca-operation-controller
      operationId: deleteOperation
      parameters:
        - name: operationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
  /api/v1/sca/operations/{operationId}/history/{historyId}:
    get:
      tags:
        - sca-operation-history-controller
      operationId: getHistory
      parameters:
        - name: operationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: historyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SCAOperationHistoryDTO'
    put:
      tags:
        - sca-operation-history-controller
      operationId: updateHistory
      parameters:
        - name: operationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: historyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SCAOperationHistoryDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SCAOperationHistoryDTO'
    delete:
      tags:
        - sca-operation-history-controller
      operationId: deleteHistory
      parameters:
        - name: operationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: historyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
...
  /api/v1/sca/operations/{operationId}/history:
    get:
      tags:
        - sca-operation-history-controller
      operationId: getAllHistory
      parameters:
        - name: operationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pageNumber
          in: query
          description: The zero-based page number to retrieve.
          required: false
          schema:
            type: string
            description: The zero-based page number to retrieve.
            example: 0
            default: '0'
          example: 0
        - name: pageSize
          in: query
          description: The number of items per page.
          required: false
          schema:
            type: string
            description: The number of items per page.
            example: 10
            default: '10'
          example: 10
        - name: sortBy
          in: query
          description: The field to sort the results by.
          required: false
          schema:
            type: string
            description: The field to sort the results by.
            example: name
          example: name
        - name: sortDirection
          in: query
          description: The direction of sorting, either ASC or DESC.
          required: false
          schema:
            type: string
            description: The direction of sorting, either ASC or DESC.
            example: DESC
            default: DESC
          example: DESC
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseSCAOperationHistoryDTO'
    post:
      tags:
        - sca-operation-history-controller
      operationId: createHistory
      parameters:
        - name: operationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SCAOperationHistoryDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SCAOperationHistoryDTO'
...
components:
  schemas:
    SCAOperationDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        referenceId:
          type: string
        operationType:
          type: string
          enum:
            - TRANSFER
            - PASSWORD_CHANGE
            - LOGIN
            - LOAN_REQUEST
            - ONBOARDING
            - OTHER
        partyId:
          type: string
        status:
          type: string
          enum:
            - PENDING
            - VERIFIED
            - FAILED
            - EXPIRED
            - CANCELLED
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        lastUpdated:
          type: string
          format: date-time
        cancelledAt:
          type: string
          format: date-time
    SCAOperationHistoryDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        scaOperationId:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - PENDING
            - VERIFIED
            - FAILED
            - EXPIRED
            - CANCELLED
        eventTime:
          type: string
          format: date-time
        comments:
          type: string
...
    PaginationResponseSCAOperationHistoryDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/SCAOperationHistoryDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...